+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/card_units.js
@@ -487,6 +487,10 @@ define([
         id: "gwaio_upgrade_kessler",
         units: [gwoUnit.kessler],
       },
+      {
+        id: "gwaio_protocol_precision",
+        units: gwoGroup.combat,
+      },
     ],
   };
 });